# api-docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Time Here Now API",
    "version": "20251023",
    "description": "API to get the current time based on timezone or client IP. **IMPORTANT: All time values are sourced from NEAR blockchain, NOT system/NTP time.** The API polls NEAR blockchain at 5 Hz (200ms intervals) and serves cached blockchain timestamps. Time zones and offsets are sourced from the IANA Time Zone Database (tzdb). Returns HTTP 503 if blockchain time is unavailable."
  },
  "servers": [
    { "url": "https://timeherenow.rodit.org:8443/api" },
    { "url": "http://localhost:8080/api" }
  ],
...
  "paths": {
    "/login": {
      "post": {
        "summary": "Authenticate and obtain session token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roditToken"],
                "properties": {
                  "roditToken": { "type": "string", "description": "RODiT authentication token" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/logout": {
      "post": {
        "summary": "Terminate session",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LogoutResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/signclient": {
      "post": {
        "summary": "Sign and mint a new RODiT client token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tobesignedValues", "mintingfee", "mintingfeeaccount"],
                "properties": {
                  "tobesignedValues": {
                    "type": "object",
                    "required": ["not_after", "max_requests", "maxrq_window", "permissioned_routes", "serviceprovider_signature"],
                    "properties": {
                      "not_after": { "type": "string", "description": "ISO 8601 timestamp for token expiration" },
                      "max_requests": { "type": "number", "description": "Maximum number of requests allowed" },
                      "maxrq_window": { "type": "number", "description": "Time window for max requests in seconds" },
                      "permissioned_routes": { "type": "string", "description": "JSON string of permitted routes and permissions" },
                      "serviceprovider_signature": { "type": "string", "description": "Service provider signature" }
                    }
                  },
                  "mintingfee": { "type": "string", "description": "Fee for minting the token" },
                  "mintingfeeaccount": { "type": "string", "description": "Account to receive the minting fee" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token successfully minted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignClientResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/timezone": {
      "post": {
        "summary": "a listing of all timezones.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "$ref": "#/components/responses/SuccessfulListJsonResponse"
          }
        }
      }
    },
    "/timezone/area": {
      "post": {
        "summary": "a listing of all timezones available for that area.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["area"],
                "properties": {
                  "area": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessfulListJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/timezone/time": {
      "post": {
        "summary": "request the current time for a timezone (falls back to user IP if timezone is omitted).",
        "description": "Returns current time for the specified timezone using NEAR blockchain time (not system/NTP time). The API polls NEAR blockchain at 5Hz and serves cached timestamps. If timezone is omitted, determines timezone from the user's IP address. Returns HTTP 503 if blockchain time is unavailable.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "timezone": { "type": "string", "description": "IANA tzdb ID, e.g., Europe/Berlin or America/Indiana/Knox. If omitted, server will determine timezone from the user's IP." },
                  "locale": { "type": "string", "description": "Optional IETF BCP 47 locale tag or Accept-Language value to localize names and datetime" },
                  "area": { "type": "string", "deprecated": true, "description": "Legacy segmented parameter for timezone area. Prefer 'timezone'." },
                  "location": { "type": "string", "deprecated": true, "description": "Legacy segmented parameter for timezone location. Prefer 'timezone'." },
                  "region": { "type": "string", "deprecated": true, "description": "Legacy segmented parameter for timezone region. Prefer 'timezone'." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessfulDateTimeJsonResponse" },
          "503": {
            "description": "NEAR time unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorJsonResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/timezones/by-country": {
      "post": {
        "summary": "list timezones by ISO 3166-1 alpha-2 country code",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["country_code"],
                "properties": {
                  "country_code": { "type": "string", "description": "ISO 3166-1 alpha-2 country code, e.g., US, DE, IN" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessfulListJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/ip": {
      "post": {
        "summary": "request the current time based on the ip of the request or provided IP (IPv4 or IPv6).",
        "description": "Returns current time for the timezone determined from the client's IP address (or provided IP) using NEAR blockchain time (not system/NTP time). Uses geoip-lite for IP-to-timezone resolution. Returns HTTP 503 if blockchain time is unavailable.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ip": { "type": "string", "description": "Optional IPv4 or IPv6 address" },
                  "locale": { "type": "string", "description": "Optional IETF BCP 47 locale tag or Accept-Language value to localize names and datetime" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SuccessfulDateTimeJsonResponse" },
          "503": {
            "description": "NEAR time unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorJsonResponse" }
              }
            }
          },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/sign/hash": {
      "post": {
        "summary": "Sign a base64url-encoded hash concatenated with the latest NEAR timestamp, likely_time_difference_ms, and public key.",
        "description": "Signs a hash with NEAR blockchain timestamp to create a tamper-proof timestamped signature. The signature covers: hash + blockchain_timestamp + likely_time_difference_ms + public_key. This proves the hash existed at or before the blockchain time. Returns HTTP 503 if blockchain time is unavailable.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["hash_b64url"],
                "properties": {
                  "hash_b64url": { "type": "string", "description": "Base64url-encoded hash. Decoded length must be 1..128 bytes." }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signature and concatenated data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SignHashResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "503": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
    "/timers/schedule": {
      "post": {
        "summary": "schedule a delayed webhook using SDK-configured destination.",
        "description": "Schedules a delayed webhook that fires after the specified delay. **All timestamps (scheduled_at, execute_at, fired_at) use NEAR blockchain time, not system time.** The webhook is sent to the destination configured in the RODiT SDK client. The timer uses blockchain time for both scheduling and firing to ensure tamper-proof timing.\n\n**Blockchain Time Granularity:** NEAR blockchain produces blocks at ~500-600ms intervals. Timestamps advance in discrete steps, not continuously. fired_at is guaranteed to be >= execute_at to maintain temporal consistency. Due to blockchain time granularity and caching, fired_at may equal execute_at even when actual delivery occurs later.\n\n**Minimum Granularity:** 1 second. Sub-second delays are not supported due to blockchain time characteristics.\n\n**Persistence:** Timers are automatically saved to disk every hour and restored on server restart. Timers that should have already fired are never sent late - they are skipped on restore. Maximum data loss window is 1 hour between auto-saves.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["delay_seconds"],
                "properties": {
                  "delay_seconds": { "type": "number", "minimum": 1, "maximum": 172800, "description": "Delay in seconds before webhook is sent (1 second to 48 hours)" },
                  "payload": { "type": "object", "description": "Optional payload echoed back in the webhook" }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Timer scheduled",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TimerScheduleResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "400": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "default": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
...
    "/mcp/resources": {
      "get": {
        "summary": "List available MCP resources",
        "description": "Returns a list of resources available through the Model Context Protocol interface",
        "tags": ["MCP"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer" },
            "description": "Maximum number of resources to return"
          },
          {
            "in": "query",
            "name": "cursor",
            "schema": { "type": "string" },
            "description": "Cursor for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of available resources with pagination info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "uri": { "type": "string" },
                          "name": { "type": "string" },
                          "type": { "type": "string" }
                        }
                      }
                    },
                    "nextCursor": { "type": "string", "nullable": true },
                    "requestId": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
...
    "/mcp/schema": {
      "get": {
        "summary": "Get MCP schema",
        "description": "Returns the OpenAPI schema for the MCP interface",
        "tags": ["MCP"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "MCP schema",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/ErrorJsonResponse" },
          "500": { "$ref": "#/components/responses/ErrorJsonResponse" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "ErrorJsonResponse": {
        "description": "an error response in JSON format",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorJsonResponse"
            }
          }
        }
      }
    },
    "schemas": {
...
      "LoginResponse": {
        "type": "object",
        "required": ["success", "message", "user", "token"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "user": {
            "type": "object",
            "required": ["id"],
            "properties": {
              "id": { "type": "string" }
            }
          },
          "token": { "type": "string", "description": "JWT authentication token" }
        }
      },
      "LogoutResponse": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" }
        }
      },
      "SignClientResponse": {
        "type": "object",
        "required": ["token_id"],
        "properties": {
          "token_id": { "type": "string", "description": "Unique identifier for the minted RODiT token" },
          "fee_signature_base64url": { "type": "string", "description": "Base64url-encoded signature for the minting fee" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}